Support clearing autofilled form inputs.
Sending a field value to empty string via setFieldValues will clear the
autofilled state for the given field.

Autofill state is also now persisted and attached to the field data object, to
be sent back to the viewer when requesting new autocompletions, allowing the
viewer to decide whether or not to render the 'clear form' button based on
whether or not there is any current persisted autofill state within the
current form.

Change-Id: Iffe17fd20af5db7fb63612c540e9e2fef68bd4ad