Add support for sign in protocol to AMP Actions
This enables the subsitution variable IDENTITY_TOKEN, which will query the
viewer with 'getAccessTokenPassive' for the identity token. Additionally, it
enables a 'amp-action.signIn' action on the amp-action element, which can be
used to trigger the 'requestSignIn' message to the viewer. If the user signs
in in the viewer, the 'signedIn' event is fired on the 'amp-action' element.

Additionally, two CSS classes (amp-action-identity-available,
amp-action-identity-unavailable) will be set on the root element of the AMP doc
indicating the signed in state if the subsitution variable / event are used.

Change-Id: Ia8d70945a791e6e358b4bca40e89860e8325554e