Made error handler in onRedner function of Google Pay inline to throw an error in order to correctly chain a promise with rejected state
The problem was when init payments client returns a rejected promise, it
will go through the rejected handler but continue with resolved state.
This is because the promise returned by then will be in rejected state
only if the handler throws an error or return a rejected promise. Since
the old handler do neither of these, the returned promise will be
resolved.

See
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/then#Return_value
for more info

Patch Set 2:
Change the error handler to use this.user().error() so that the error
can be tracked

Patch Set 3:
Error handler now reject with the error object other than the error
message

Change-Id: I112b453ba588af452f5f3c9b7d5d31719e6a9453