Fixed two issues that causes our unit test for Gpay button and Gpay inline widget to break.
One issue is caused by commit
https://team.git.corp.google.com/aog-amp-actions/amphtml/+/115d97c5355cfb48dcd786c13760b62caff0303f.
It addes a new method canRenderTemplates to viewer and amp form calls
that function. Therefore we need to stub it in our unit test.

The other issue is caused by commit
https://team.git.corp.google.com/aog-amp-actions/amphtml/+/c7e5fc3cd01529a7ac2f64ddaec014ccc9b0c73b.
It forbids the import of sinon. Importing sinon breaks our stubs in unit
tests and everything works correctly if we don't import sinon

Also added a expect error to the unit test that should have error

Change-Id: I6b4bad3e7886d65bf2f0c52cb7c3b176b294157e